+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -304,7 +304,6 @@ instance Arbitrary FreeVars where
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
-      <*> arbitrary
   shrink = genericShrink
 
 ------------------------------------------------------------------------------------------